requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorzero==2.0
exceptiongroup==1.2.1
gpiozero==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==4.13.0
importlib-resources==5.13.0
notion-client==2.2.1
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
schedule==1.2.1
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
